requirements.txt
# requirements.txt
openai-whisper
transformers
gTTS
